refactor: navigate to correct stack after login
