+++ src/navigation/TabNavigator.tsx
@@ -20,20 +20,13 @@ import {
 } from '@atb/utils/navigation';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import {LabelPosition} from '@react-navigation/bottom-tabs/lib/typescript/src/types';
-import {
-  NavigatorScreenParams,
-  ParamListBase,
-  useNavigation,
-} from '@react-navigation/native';
-import React, {useEffect} from 'react';
+import {NavigatorScreenParams, ParamListBase} from '@react-navigation/native';
+import React from 'react';
 import {SvgProps} from 'react-native-svg';
 import ThemeIcon from '@atb/components/theme-icon/theme-icon';
 import DeparturesScreen from '@atb/screens/Departures';
 import {TicketTabsNavigatorParams} from '@atb/screens/Ticketing/Tickets';
-import {useHasEnabledMobileToken} from '@atb/mobile-token/MobileTokenContext';
-import {useAuthState} from '@atb/auth';
-import {useAppState} from '@atb/AppContext';
-import {shouldOnboardMobileToken} from '@atb/api/utils';
+import {useGoToMobileTokenOnboardingWhenNecessary} from '@atb/screens/MobileTokenOnboarding/utils';
 
 type SubNavigator<T extends ParamListBase> = {
   [K in keyof T]: {screen: K; initial?: boolean; params?: T[K]};
@@ -156,21 +149,3 @@ function tabSettings(
     tabBarIcon: ({color}) => <ThemeIcon svg={Icon} fill={color} />,
   };
 }
-
-const useGoToMobileTokenOnboardingWhenNecessary = () => {
-  const hasEnabledMobileToken = useHasEnabledMobileToken();
-  const {authenticationType} = useAuthState();
-  const {mobileTokenOnboarded} = useAppState();
-  const shouldOnboard = shouldOnboardMobileToken(
-    hasEnabledMobileToken,
-    authenticationType,
-    mobileTokenOnboarded,
-  );
-  const navigation = useNavigation();
-
-  useEffect(() => {
-    if (shouldOnboard) {
-      navigation.navigate('MobileTokenOnboarding');
-    }
-  }, [shouldOnboard]);
-};
