+++ src/screens/Profile/Home/index.tsx
@@ -160,7 +160,10 @@ export default function ProfileHome({navigation}: ProfileScreenProps) {
                       ? 'LoginOptionsScreen'
                       : 'PhoneInputInApp',
                     params: {
-                      afterLogin: {routeName: 'ProfileHome'},
+                      afterLogin: {
+                        routeName: 'TabNavigator',
+                        routeParams: {screen: 'Profile'},
+                      },
                     },
                   })
                 }
